:running: moved test project v2 dir
Moving testdata/project_v2 --> testdata/project-v2 to ensure
project-v2 project can be deployed on a kubernetes cluster.
 - project-name is used as namespace name and name prefix and dash '-' character is
not allowed in namespace and nameprefix in kubernetes world.
