+++ testdata/project-v2/api/v1/groupversion_info.go
