+++ testdata/project-v2/config/certmanager/certificate.yaml
