+++ testdata/project-v2/config/certmanager/kustomization.yaml
