+++ testdata/project-v2/config/certmanager/kustomizeconfig.yaml
