+++ testdata/project-v2/config/crd/bases/crew.testproject.org_captains.yaml
