+++ testdata/project-v2/config/crd/kustomization.yaml
