+++ testdata/project-v2/config/crd/kustomizeconfig.yaml
