+++ testdata/project-v2/config/crd/patches/webhook_in_captain.yaml
