+++ testdata/project-v2/config/default/manager_auth_proxy_patch.yaml
