+++ testdata/project-v2/config/default/manager_image_patch.yaml
