+++ testdata/project-v2/config/default/manager_prometheus_metrics_patch.yaml
