+++ testdata/project-v2/config/manager/kustomization.yaml
