+++ testdata/project-v2/config/rbac/auth_proxy_role_binding.yaml
