+++ testdata/project-v2/config/rbac/auth_proxy_service.yaml
