+++ testdata/project-v2/config/rbac/kustomization.yaml
