+++ testdata/project-v2/config/samples/crew_v1_captain.yaml
