+++ testdata/project-v2/go.mod
@@ -1,4 +1,4 @@
-module sigs.k8s.io/kubebuilder/testdata/project_v2
+module sigs.k8s.io/kubebuilder/testdata/project-v2
 
 go 1.12
 
@@ -8,5 +8,4 @@ require (
 	k8s.io/apimachinery v0.0.0-20190221213512-86fb29eff628
 	k8s.io/client-go v0.0.0-20190228174230-b40b2a5939e4
 	sigs.k8s.io/controller-runtime v0.2.0-alpha.0.0.20190503051552-b666157c41da
-	sigs.k8s.io/controller-tools v0.2.0-alpha.1 // indirect
 )
