--- testdata/project_v2/PROJECT
@@ -1,3 +1,0 @@
-version: "2"
-domain: testproject.org
-repo: sigs.k8s.io/kubebuilder/testdata/project_v2
